refactor: code improvements
