+++ spec/providers/firestore.spec.ts
@@ -1,6 +1,6 @@
 import { expect } from 'chai';
 import * as firebase from 'firebase-admin';
-import { FeaturesList } from '../../src/features';
+import { FeaturesList } from '../../src/types/commonTypes';
 import fft = require('../../src/index');
 
 describe('providers/firestore', () => {
