+++ src/cloudevent/generate.ts
@@ -8,10 +8,11 @@ import {
   DocumentSnapshot,
   QueryDocumentSnapshot,
 } from 'firebase-admin/firestore';
-import { LIST_OF_MOCK_CLOUD_EVENT_PARTIALS } from './mocks/partials';
-import { DeepPartial } from './types';
 import { Change } from 'firebase-functions/v1';
 import merge from 'ts-deepmerge';
+
+import { LIST_OF_MOCK_CLOUD_EVENT_PARTIALS } from './mocks/partials';
+import { DeepPartial } from './types';
 
 /**
  * @return {CloudEvent} Generated Mock CloudEvent
